refactor monolith
