--- api/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "prodctrl-api"
-description = "production control business api"
-authors.workspace = true
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[features]
-default = []
-db_compat = [ "tiberius" ]
-
-[dependencies]
-serde = { workspace = true }
-regex = { workspace = true }
-lazy_static = { workspace = true }
-tiberius = { version = "0.11.0", optional = true }
