--- config/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "prodctrl-config"
-description = "production control embedded configurations"
-authors.workspace = true
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[dependencies]
-rust-embed = { version = "6.4.1", features = ["debug-embed", "include-exclude"] }
-serde = { workspace = true }
-toml = "0.5.9"
