--- config/build.rs
@@ -1,5 +1,0 @@
-fn main() -> std::io::Result<()> {
-    println!("cargo:warning=⚙️ Config build script OK");
-
-    Ok(())
-}
