--- db/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "prodctrl-db"
-description = "production control database connectors"
-authors.workspace = true
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[dependencies]
-prodctrl-api = { workspace = true, features = ["db_compat"] }
-prodctrl-config = { workspace = true }
-serde = { workspace = true }
-log = "0.4.17"
-
-# async
-async_once = "0.2"
-async-trait = "0.1"
-# crossbeam = "0.8"
-
-# MSSQL database
-bb8 = "0.8"
-# see tokio-util dependency note below before upgrading bb8-tiberius version
-bb8-tiberius = "0.12.0"
-
-[dependencies.tiberius]
-version = "0.10"
-features = [ "sql-browser-tokio", "integrated-auth-gssapi" ]
-
-[dependencies.tokio]
-version = "1.19"
-features = ["net", "rt-multi-thread", "sync"]
-
-[dependencies.tokio-util]
-# tokio-util version must match version used in bb8-tiberius for type compatibility
-version = "0.6.2"
-features = [ "compat" ]
